+++ sharkfuser/include/fusilli.h
@@ -17,7 +17,7 @@
 #include "fusilli/asm_emitter.h"
 #include "fusilli/external_tools.h"
 #include "fusilli/extras.h"
-#include "fusilli/logging.h"
+#include "fusilli/support/logging.h"
 #include "fusilli/types.h"
 
 // Attributes
