+++ sharkfuser/include/fusilli/attributes/tensor_attributes.h
@@ -15,7 +15,7 @@
 #define FUSILLI_ATTRIBUTES_TENSOR_ATTRIBUTES_H
 
 #include "fusilli/graph/context.h"
-#include "fusilli/logging.h"
+#include "fusilli/support/logging.h"
 #include "fusilli/types.h"
 
 #include <algorithm>
